Added skeleton for IRC bot
bot.py is the main body of the bot. 
IRC_bot.py is the class that holds the necessary functions. 
Program is runnable with Python3 but will fail with Python2 due to changes in the way UTF-8 encoding is handled. 

This bot connects to the specified IRC channel on free node and sends a hello world message. 
It then parses whatever is typed into the chat and prints it locally.